requirements.txt
streamlit
pandas
plotly
requests
python-dotenv
audio-recorder-streamlit